Domain separation (#15)
* fix major bug with PBC on in positions.hpp

* remove empty utils.hpp file

* add variable smoothing length

* fix bug in Printer with dt[0] when node has 0 particles

* add maxvsignal for timestep calculation

* use smaller global tree for smaller test cases

* remove old comments

* remove JENKINS in sqpatch.cpp

* renamed timer.hpp and removed unused variable

* split DistributedDomain into Domain and DistributedDomain

* add findNeighbors and updateSmoothinglength module (from Domain)

* add missing files

* use unordered_map everywhere

* fix include problem in omp only version

* address review comments
