+++ src/include/DistributedDomain.hpp
@@ -1,139 +1,42 @@
 #pragma once
 
+#include <unordered_map>
+
+#include "Domain.hpp"
+
 #ifdef USE_MPI
 #include "mpi.h"
 #endif
 
-#include <vector>
-#include <cmath>
-#include <unordered_map>
-#include <map>
-#include <algorithm>
-#include <numeric>
-
-#include <unistd.h>
-#include <chrono>
-#include <ctime>
-
-#include "Octree.hpp"
-#include "Task.hpp"
-
 namespace sphexa
 {
 
-template <typename T>
-class DistributedDomain
+template <typename T, class Dataset>
+class DistributedDomain : public Domain<T, Dataset>
 {
 public:
     DistributedDomain()
     {
-#ifdef USE_MPI
         MPI_Comm_size(MPI_COMM_WORLD, &comm_size);
         MPI_Comm_rank(MPI_COMM_WORLD, &comm_rank);
         MPI_Get_processor_name(processor_name, &name_len);
-#endif
     }
 
-    template <class Dataset>
-    void findNeighborsImpl(Task &t, Dataset &d)
-    {
-        size_t  n = t.clist.size();
+    ~DistributedDomain() override = default;
 
-#pragma omp parallel for schedule(guided)
-        for (size_t pi = 0; pi < n; pi++)
-        {
-            int i = t.clist[pi];
-
-            t.neighborsCount[pi] = 0;
-            octree.findNeighbors(i, &d.x[0], &d.y[0], &d.z[0], d.x[i], d.y[i], d.z[i], 2.0 * d.h[i], t.ngmax, &t.neighbors[pi * t.ngmax],
-                                 t.neighborsCount[pi], d.bbox.PBCx, d.bbox.PBCy, d.bbox.PBCz);
-
-#ifndef NDEBUG
-            if (t.neighborsCount[pi] == 0)
-                printf("ERROR::FindNeighbors(%d) x %f y %f z %f h = %f ngi %d\n", i, d.x[i], d.y[i], d.z[i], d.h[i], t.neighborsCount[pi]);
-#endif
-        }
-    }
-
-    template <class Dataset>
-    void findNeighbors(std::vector<Task> &taskList, Dataset &d)
-    {
-        for (auto &task : taskList)
-        {
-            findNeighborsImpl(task, d);
-        }
-    }
-
-    size_t  neighborsSumImpl(const Task &t)
-    {
-        size_t  sum = 0;
-#pragma omp parallel for reduction(+ : sum)
-        for (unsigned int i = 0; i < t.clist.size(); i++)
-            sum += t.neighborsCount[i];
-
-        return sum;
-    }
-
-    size_t  neighborsSum(const std::vector<Task> &taskList)
-    {
-        size_t  sum = 0;
-        for (const auto &task : taskList)
-        {
-            sum += neighborsSumImpl(task);
-        }
-
-#ifdef USE_MPI
-        MPI_Allreduce(MPI_IN_PLACE, &sum, 1, MPI_LONG_LONG_INT, MPI_SUM, MPI_COMM_WORLD);
-#endif
-
-        return sum;
-    }
-
-    inline T normalize(T d, T min, T max) { return (d - min) / (max - min); }
-
-    void reorderSwap(const std::vector<int> &ordering, std::vector<T> &arrayList)
-    {
-        std::vector<T> tmp(ordering.size());
-        for (unsigned int i = 0; i < ordering.size(); i++)
-            tmp[i] = arrayList[ordering[i]];
-        tmp.swap(arrayList);
-    }
-
-    void reorder(const std::vector<int> &ordering, std::vector<std::vector<T> *> &arrayList)
-    {
-        for (unsigned int i = 0; i < arrayList.size(); i++)
-            reorderSwap(ordering, *arrayList[i]);
-    }
-
-    template <class Dataset>
-    void reorder(const std::vector<int> &ordering, Dataset &d)
-    {
-        reorder(ordering, d.data);
-    }
-
-    void makeDataArray(std::vector<std::vector<T> *> &data, std::vector<T> *d) { data.push_back(d); }
-
-    template <typename... Args>
-    void makeDataArray(std::vector<std::vector<T> *> &data, std::vector<T> *first, Args... args)
-    {
-        data.push_back(first);
-        makeDataArray(data, args...);
-    }
-
-#ifdef USE_MPI
     /*  Send the particles from the old arrays either to the new arrays or to another process
         Basically collects a map[process][nodes] to send to other processes
         Then it knows how many particle it is missing (particleCount - localParticleCount)
         And just loop receive until we have received all the missing particles from other processes */
     void sync(const std::vector<std::vector<T> *> &arrayList)
     {
-        std::map<int, std::vector<int>> toSendCellsPadding, toSendCellsCount;
+        std::unordered_map<int, std::vector<int>> toSendCellsPadding, toSendCellsCount;
         std::vector<std::vector<T>> buff;
         std::vector<MPI_Request> requests;
 
         int needed = 0;
 
-        octree.syncRec(toSendCellsPadding, toSendCellsCount, needed);
+        Domain<T, Dataset>::octree.syncRec(toSendCellsPadding, toSendCellsCount, needed);
 
         for (int rank = 0; rank < comm_size; rank++)
         {
@@ -158,7 +61,8 @@ public:
                         int padding = toSendCellsPadding[rank][j];
                         int count = toSendCellsCount[rank][j];
 
-                        std::copy((*arrayList[i]).begin() + padding, (*arrayList[i]).begin() + padding + count, std::back_inserter(localBuffer));
+                        std::copy((*arrayList[i]).begin() + padding, (*arrayList[i]).begin() + padding + count,
+                                  std::back_inserter(localBuffer));
                     }
 
                     MPI_Isend(localBuffer.data(), nParticlesToSend, MPI_DOUBLE, rank, 2 + i, MPI_COMM_WORLD, &requests[rcount + 2 + i]);
@@ -211,7 +115,7 @@ public:
         // typedef std::chrono::time_point<Clock> TimePoint;
         // typedef std::chrono::duration<float> Time;
 
-        std::map<int, Octree<T> *> cellMap;
+        std::unordered_map<int, Octree<T> *> cellMap;
 
         struct ToSend
         {
@@ -458,8 +362,7 @@ public:
         }
     }
 
-    template <class Dataset>
-    void create(Dataset &d)
+    void create(Dataset &d) override
     {
         const std::vector<T> &x = d.x;
         const std::vector<T> &y = d.y;
@@ -472,11 +375,17 @@ public:
         const size_t split = ntot / comm_size;
         const size_t remaining = ntot - comm_size * split;
 
+        std::vector<int> &clist = Domain<T, Dataset>::clist;
+
         clist.resize(n);
+
+        #pragma omp parallel for
         for (size_t i = 0; i < n; i++)
             clist[i] = i;
 
         d.bbox.computeGlobal(clist, d.x, d.y, d.z);
+
+        Octree<T> &octree = Domain<T, Dataset>::octree;
 
         octree.cells.clear();
         octree = Octree<T>(d.bbox.xmin, d.bbox.xmax, d.bbox.ymin, d.bbox.ymax, d.bbox.zmin, d.bbox.zmax, comm_rank, comm_size);
@@ -511,7 +420,7 @@ public:
         // adjust clist to consider only the particles that belong to us (using the tree)
         clist.resize(workAssigned);
         ordering.resize(workAssigned);
-        buildTree(d);
+        Domain<T, Dataset>::buildTree(d);
 
         // build the global tree using only the particles that belong to us
         octree.globalRebalance(d.bbox.xmin, d.bbox.xmax, d.bbox.ymin, d.bbox.ymax, d.bbox.zmin, d.bbox.zmax);
@@ -520,8 +429,11 @@ public:
 
         // Get rid of particles that do not belong to us
         reorder(ordering, d);
+
+        #pragma omp parallel for
         for (size_t i = 0; i < workAssigned; i++)
             clist[i] = i;
+        
         d.resize(workAssigned);
 
         workAssigned = work[comm_rank] - work_remaining[comm_rank];
@@ -531,8 +443,7 @@ public:
         d.count = workAssigned;
     }
 
-    template <class Dataset>
-    void distribute(Dataset &d)
+    void update(Dataset &d) override
     {
         const std::vector<T> &x = d.x;
         const std::vector<T> &y = d.y;
@@ -544,6 +455,9 @@ public:
         const size_t ntot = d.n;
         const size_t split = ntot / comm_size;
         const size_t remaining = ntot - comm_size * split;
+
+        std::vector<int> &clist = Domain<T, Dataset>::clist;
+        Octree<T> &octree = Domain<T, Dataset>::octree;
 
         d.bbox.computeGlobal(clist, d.x, d.y, d.z);
 
@@ -619,137 +533,14 @@ public:
 
         d.count = workAssigned;
     }
-#else
-    template <class Dataset>
-    void create(Dataset &d)
-    {
-        const std::vector<T> &x = d.x;
-        const std::vector<T> &y = d.y;
-        const std::vector<T> &z = d.z;
-
-        const size_t n = d.count;
-
-        clist.resize(n);
-        for (size_t i = 0; i < n; i++)
-            clist[i] = i;
-
-        std::vector<int> ordering(n);
-
-        d.bbox.computeGlobal(clist, d.x, d.y, d.z);
-
-        // Each process creates a tree based on the gathered sample
-        octree.cells.clear();
-        octree = Octree<T>(d.bbox.xmin, d.bbox.xmax, d.bbox.ymin, d.bbox.ymax, d.bbox.zmin, d.bbox.zmax, 0, 1);
-        octree.buildTree(clist, x, y, z, ordering);
-        reorder(ordering, d);
-    }
-
-    template <class Dataset>
-    void distribute(Dataset &d)
-    {
-        d.bbox.computeGlobal(clist, d.x, d.y, d.z);
-        octree.cells.clear();
-        octree.xmin = d.bbox.xmin;
-        octree.xmax = d.bbox.xmax;
-        octree.ymin = d.bbox.ymin;
-        octree.ymax = d.bbox.ymax;
-        octree.zmin = d.bbox.zmin;
-        octree.zmax = d.bbox.zmax;
-        return;
-    }
-    template <typename... Args>
-    void synchronizeHalos(Args...)
-    {
-        return;
-    }
-
-    void synchronizeHalos(std::vector<std::vector<T> *> &) { return; }
-#endif
-
-    template <class Dataset>
-    void buildTree(Dataset &d)
-    {
-        // Finally remap everything
-        std::vector<int> ordering(d.x.size());
-
-        std::vector<int> list(d.x.size());
-#pragma omp parallel for
-        for (int i = 0; i < (int)d.x.size(); i++)
-            list[i] = i;
-
-        // We need this to expand halo
-        octree.buildTree(list, d.x, d.y, d.z, ordering);
-        reorder(ordering, d);
-
-        octree.mapList(clist);
-    }
-
-    template <class Dataset>
-    void updateSmoothingLengthImpl(Task &t, Dataset &d)
-    {
-        const T c0 = 7.0;
-        const T exp = 1.0 / 3.0;
-        
-        const int ng0 = Task::ng0;
-        const int *neighborsCount = t.neighborsCount.data();
-        T *h = d.h.data();
-
-        size_t  n = t.clist.size();
-
-#pragma omp parallel for schedule(guided)
-        for (size_t pi = 0; pi < n; pi++)
-        {
-            int i = t.clist[pi];
-            const int nn = neighborsCount[pi];
-
-            h[i] = h[i] * 0.5 * pow((1.0 + c0 * ng0 / nn), exp);
-
-#ifndef NDEBUG
-            if (std::isinf(h[i]) || std::isnan(h[i])) printf("ERROR::h(%d) ngi %d h %f\n", i, nn, h[i]);
-#endif
-        }
-    }
-
-    template <class Dataset>
-    void updateSmoothingLength(std::vector<Task> &taskList, Dataset &d)
-    {
-        for (auto &task : taskList)
-        {
-            updateSmoothingLengthImpl(task, d);
-        }
-    }
-
-    void createTasks(std::vector<Task> &taskList, const size_t nTasks)
-    {
-        const int partitionSize = clist.size() / nTasks;
-        const int lastPartitionOffset = clist.size() - nTasks * partitionSize;
-
-        taskList.resize(nTasks);
-
-#pragma omp parallel for
-        for (size_t i = 0; i < nTasks; ++i)
-        {
-            const int begin = i * partitionSize;
-            const int end = (i + 1) * partitionSize + (i == nTasks - 1 ? lastPartitionOffset : 0);
-            const size_t size = end - begin;
-
-            taskList[i].resize(size);
-            for (size_t j = 0; j < size; j++)
-                taskList[i].clist[j] = clist[j + begin];
-        }
-    }
 
     int comm_size, comm_rank, name_len;
     char processor_name[256];
 
-    std::map<int, std::map<int, Octree<T> *>> toSendHalos;
+    std::unordered_map<int, std::unordered_map<int, Octree<T> *>> toSendHalos;
 
     size_t haloCount = 0;
     size_t workAssigned = 0;
-
-    Octree<T> octree;
-
-    std::vector<int> clist;
 };
 
-} // namespace sphexa
+}
