+++ src/include/sphexa.hpp
@@ -2,13 +2,13 @@
 
 #ifdef USE_MPI
 #include "mpi.h"
+#include "DistributedDomain.hpp"
 #endif
 
-#include "DistributedDomain.hpp"
+#include "Domain.hpp"
 #include "Octree.hpp"
 #include "BBox.hpp"
 #include "Task.hpp"
-
 #include "ArgParser.hpp"
 #include "config.hpp"
 #include "Timer.hpp"
@@ -18,6 +18,7 @@
 #include "sph/cuda/sph.cuh"
 #endif
 
+#include "sph/findNeighbors.hpp"
 #include "sph/density.hpp"
 #include "sph/IAD.hpp"
 #include "sph/momentumAndEnergyIAD.hpp"
@@ -26,3 +27,4 @@
 #include "sph/timestep.hpp"
 #include "sph/positions.hpp"
 #include "sph/totalEnergy.hpp"
+#include "sph/updateSmoothingLength.hpp"
