+++ src/sqpatch.cpp
@@ -21,43 +21,45 @@ int main(int argc, char **argv)
 
 #ifdef USE_MPI
     MPI_Init(NULL, NULL);
+    DistributedDomain<Real, Dataset> domain;
+#else
+    Domain<Real, Dataset> domain;
 #endif
 
     auto d = SqPatchDataGenerator<Real>::generate(cubeSide);
-    DistributedDomain<Real> distributedDomain;
     Printer<Dataset> printer(d);
-    MPITimer timer(d.rank);
+    MasterProcessTimer timer(d.rank);
 
     std::ofstream constantsFile("constants.txt");
 
     std::vector<Task> taskList;
 
-    distributedDomain.create(d);
+    domain.create(d);
 
     for (d.iteration = 0; d.iteration <= maxStep; d.iteration++)
     {
         timer.start();
 
-        distributedDomain.distribute(d);
+        domain.update(d);
         timer.step("domain::distribute");
-        distributedDomain.synchronizeHalos(&d.x, &d.y, &d.z, &d.h);
+        domain.synchronizeHalos(&d.x, &d.y, &d.z, &d.h);
         timer.step("mpi::synchronizeHalos");
-        distributedDomain.buildTree(d);
+        domain.buildTree(d);
         timer.step("domain::buildTree");
-        distributedDomain.createTasks(taskList, 64);
+        domain.createTasks(taskList, 64);
         timer.step("domain::createTasks");
-        distributedDomain.findNeighbors(taskList, d);
+        sph::findNeighbors(domain.octree, taskList, d);
         timer.step("FindNeighbors");
         sph::computeDensity<Real>(taskList, d);
         if (d.iteration == 0) { sph::initFluidDensityAtRest<Real>(taskList, d); }
         timer.step("Density");
         sph::computeEquationOfState<Real>(taskList, d);
         timer.step("EquationOfState");
-        distributedDomain.synchronizeHalos(&d.vx, &d.vy, &d.vz, &d.ro, &d.p, &d.c);
+        domain.synchronizeHalos(&d.vx, &d.vy, &d.vz, &d.ro, &d.p, &d.c);
         timer.step("mpi::synchronizeHalos");
         sph::computeIAD<Real>(taskList, d);
         timer.step("IAD");
-        distributedDomain.synchronizeHalos(&d.c11, &d.c12, &d.c13, &d.c22, &d.c23, &d.c33);
+        domain.synchronizeHalos(&d.c11, &d.c12, &d.c13, &d.c22, &d.c23, &d.c33);
         timer.step("mpi::synchronizeHalos");
         sph::computeMomentumAndEnergyIAD<Real>(taskList, d);
         timer.step("MomentumEnergyIAD");
@@ -67,20 +69,19 @@ int main(int argc, char **argv)
         timer.step("UpdateQuantities");
         sph::computeTotalEnergy<Real>(taskList, d);
         timer.step("EnergyConservation"); // AllReduce(sum:ecin,ein)
-        distributedDomain.updateSmoothingLength(taskList, d);
+        sph::updateSmoothingLength<Real>(taskList, d);
         timer.step("UpdateSmoothingLength");
 
-        long long int totalNeighbors = distributedDomain.neighborsSum(taskList);
+        size_t totalNeighbors = sph::neighborsSum(taskList);
         if (d.rank == 0)
         {
-            printer.printCheck(distributedDomain.clist.size(), distributedDomain.octree.globalNodeCount, distributedDomain.haloCount,
-                               totalNeighbors, std::cout);
+            printer.printCheck(d.count, domain.octree.globalNodeCount, d.x.size() - d.count, totalNeighbors, std::cout);
             printer.printConstants(d.iteration, totalNeighbors, constantsFile);
         }
 
         if ((writeFrequency > 0 && d.iteration % writeFrequency == 0) || writeFrequency == 0)
         {
-            printer.printAllDataToFile(distributedDomain.clist, "dump" + to_string(d.iteration) + ".txt");
+            printer.printAllDataToFile(domain.clist, "dump" + to_string(d.iteration) + ".txt");
             timer.step("writeFile");
         }
